# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Agent-API
  version: "0.1.0"
  description: >
    Generischer API-Layer für Agenten (Mensch und KI) auf einen Baum aus
    Markdown-/JSON-/PII-JSON-/Binärdateien. Eine Instanz verwaltet genau
    einen Baum (kein Multi-Tenant-Modell, siehe spec.md §12). Grundlage:
    docs/spec.md.
  license:
    name: proprietary
servers:
  - url: /api/v1
    description: Basis-Pfad. Kein Tenant-Segment — eine Instanz je Baum/Projekt (spec.md §12).
security:
  - clientApiKey: []
    userToken: []
  - clientApiKey: []
...
paths:
  /tree:
    get:
      operationId: getTree
      tags: [Tree]
      summary: Teilbaum auflisten
      description: >
        ACL-gefiltert (spec.md §8): Einträge ohne Read-Recht für den
        aufrufenden (user_id, client_id) werden vollständig ausgeblendet,
        nicht nur inhaltlich gesperrt.
      parameters:
        - name: root
          in: query
          required: false
          schema: { type: string, default: "/" }
          description: Nur diesen Teilbaum liefern.
        - name: depth
          in: query
          required: false
          schema: { type: integer, minimum: 0, default: 2 }
          description: "Ebenen unterhalb von root. 0 = nur Metadaten von root, 1 = nur direkte Kinder, ..."
        - name: cursor
          in: query
          required: false
          schema: { type: string }
          description: Pagination-Token einer vorherigen Antwort (pro Verzeichnisebene).
      responses:
        "200":
          description: Teilbaum-Listing
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TreeResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  "/file/{path}":
    parameters:
      - $ref: "#/components/parameters/FilePath"
    get:
      operationId: getFile
      tags: [Tree]
      summary: Datei lesen
      description: >
        Text-Kinds (md/json/pii.json): Volltext + Version im Response-Body.
        `binary`: nur Metadaten, keine Bytes — dafür GET .../content (spec.md §4).
      responses:
        "200":
          description: Metadaten (+ Inhalt bei Text-Kinds)
          headers:
            ETag: { schema: { type: string }, description: "= version, für If-None-Match-Caching" }
          content:
            application/json:
              schema: { $ref: "#/components/schemas/FileResponse" }
        "304": { description: "Not Modified (If-None-Match traf zu)" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
      parameters:
        - $ref: "#/components/parameters/IfNoneMatch"
    post:
      operationId: writeFile
      tags: [Files]
      summary: Datei vollständig schreiben (Full Write / Upload)
      description: >
        Neue Datei anlegen oder vollständig ersetzen. Text-Kinds über
        `application/json`, `binary` über `multipart/form-data` (spec.md §5).
        `if_version` fehlt/leer bei Neuanlage, sonst Pflicht.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/FullWriteTextRequest" }
          multipart/form-data:
            schema: { $ref: "#/components/schemas/FullWriteBinaryRequest" }
      responses:
        "200":
          description: Geschrieben
          content:
            application/json:
              schema: { $ref: "#/components/schemas/WriteResult" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/VersionConflict" }
    delete:
      operationId: deleteFile
      tags: [Files]
      summary: Datei löschen
      description: Immer explizit, immer mit `if_version` (spec.md §5).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeleteRequest" }
      responses:
        "204": { description: Gelöscht }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/VersionConflict" }
  "/file/{path}/content":
    parameters:
      - $ref: "#/components/parameters/FilePath"
    get:
      operationId: getFileContent
      tags: [Tree]
      summary: Rohbytes einer Binärdatei herunterladen
      description: "Nur für `kind: binary`. Keine Text-Extraktion (spec.md §4) — bleibt außerhalb dieser API."
      responses:
        "200":
          description: Rohbytes
          headers:
            ETag: { schema: { type: string } }
          content:
            application/octet-stream:
              schema: { type: string, contentEncoding: base64 }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: "kind ist kein `binary` (falscher Endpoint für diesen Dateityp)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
components:
  securitySchemes:
    clientApiKey:
      type: http
      scheme: bearer
      description: >
        `Authorization: Bearer <client_api_key>` — beweist den client_id
        (spec.md §10a). Auf JEDEM Call erforderlich.
    userToken:
      type: apiKey
      in: header
      name: X-User-Token
      description: >
        Signierter Kurzzeit-JWT (EdDSA, ~15 min gültig) — beweist den
        user_id (spec.md §10b). Erforderlich bei jedem Call, AUSSER der
        Client hat `fixed_user_id` in der Registry gesetzt (autonomer
        Agent ohne Menschen dahinter) — dieser Fall ist in `security:`
        oben als zweite, alternative Anforderung (nur clientApiKey)
        hinterlegt. Welcher der beiden Fälle zutrifft, weiß der Server
        erst zur Laufzeit aus der Client-Registry — bei fälschlich
        fehlendem Token für einen nicht-autonomen Client lehnt er mit
        401 ab.
  parameters:
    FilePath:
      name: path
      in: path
      required: true
      schema: { type: string }
      description: >
        Relativer POSIX-Pfad zur Datei INKLUSIVE Schrägstrichen, z.B.
        `lv/sanitaer.md`. Wird server-seitig als Greedy-Match behandelt
        (FastAPI `path`-Konverter) — Clients dürfen den Pfad nicht
        segmentweise urlencodieren.
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      schema: { type: string }
      description: Für Caching — entspricht der zuletzt gesehenen `version`.
  schemas:
    Kind:
      type: string
      enum: [dir, md, json, "pii.json", binary]
    TreeEntry:
      type: object
      required: [path, kind]
      properties:
        path: { type: string }
        kind: { $ref: "#/components/schemas/Kind" }
        size: { type: integer, description: "Bytes, fehlt bei kind=dir" }
        version: { type: string, description: "fehlt bei kind=dir" }
        preview: { type: string, description: "nur Text-Kinds" }
        mime_type: { type: string, description: "nur kind=binary" }
    TreeResponse:
      type: object
      required: [entries]
      properties:
        entries:
          type: array
          items: { $ref: "#/components/schemas/TreeEntry" }
        next_cursor: { type: string }
    FileResponse:
      type: object
      required: [path, kind, version]
      properties:
        path: { type: string }
        kind: { $ref: "#/components/schemas/Kind" }
        version: { type: string }
        size: { type: integer }
        mime_type: { type: string, description: "nur kind=binary" }
        content: { type: string, description: "nur Text-Kinds, Volltext" }
    WriteResult:
      type: object
      required: [path, version, commit_id]
      properties:
        path: { type: string }
        version: { type: string, description: "neue Version nach dem Schreiben" }
        commit_id: { type: string, description: "Git-Commit-SHA dieser Transaktion" }
...
    FullWriteTextRequest:
      type: object
      required: [content, reason]
      properties:
        content: { type: string }
        if_version: { type: string, description: "weglassen bei Neuanlage" }
        reason: { type: string, minLength: 1 }
        task_id: { type: string }
    FullWriteBinaryRequest:
      type: object
      required: [file, reason]
      properties:
        file: { type: string, contentEncoding: binary, description: "Rohbytes, multipart" }
        if_version: { type: string, description: "= Hash der bisherigen Bytes; weglassen bei Neuanlage" }
        reason: { type: string, minLength: 1 }
        task_id: { type: string }
    DeleteRequest:
      type: object
      required: [if_version, reason]
      properties:
        if_version: { type: string }
        reason: { type: string, minLength: 1 }
        task_id: { type: string }
...
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string, description: "Maschinenlesbarer Code, z.B. unauthorized, acl_denied, not_found, no_match, ambiguous_match" }
        message: { type: string }
        details: { type: object }
    ConflictError:
      type: object
      required: [error, current_version]
      properties:
        error: { type: string, const: version_conflict }
        message: { type: string }
        current_content: { type: string, nullable: true, description: "null bei kind=binary" }
        current_version: { type: string }

    # --- Client-Registry (spec.md §10, Design-Doc §14/§15) ---
...
  responses:
    Unauthorized:
      description: "Client-API-Key fehlt/ungültig, oder User-Token fehlt/ungültig/abgelaufen wo nötig"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: "ACL verweigert (spec.md §8) oder Admin-Endpunkt ohne Admin-user_id"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Pfad/Ressource nicht gefunden, oder ACL blendet aus (Sichtbarkeit wie „nicht gefunden")
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: "Ungültige Anfrage (z.B. old_str 0 oder >1 Treffer, Pflichtfeld fehlt)"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    VersionConflict:
      description: "if_version stimmt nicht mehr mit dem aktuellen Stand überein (spec.md §11)"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ConflictError" }